feat(front): confirmação pra delete

+++ src/app/components/task-list/task-list.component.ts
@@ -12,8 +12,11 @@ import { TaskList } from '../../types/tasklist';
 import { TaskComponent } from '../task/task.component';
 import { CommonModule } from '@angular/common';
 import { MenuModule } from 'primeng/menu';
-import { MenuItem } from 'primeng/api';
+import { ConfirmationService, MenuItem, MessageService } from 'primeng/api';
 import { ListService } from '../../services/list.service';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
+import { ToastModule } from 'primeng/toast';
+import { ButtonModule } from 'primeng/button';
 
 @Component({
   selector: 'app-task-list',
@@ -26,9 +29,13 @@ import { ListService } from '../../services/list.service';
     TaskComponent,
     CommonModule,
     MenuModule,
+    ConfirmDialogModule,
+    ToastModule,
+    ButtonModule
   ],
   templateUrl: './task-list.component.html',
   styleUrl: './task-list.component.scss',
+  providers: [ConfirmationService, MessageService],
 })
 export class TaskListComponent {
   readonly minimize = Minimize2;
@@ -41,7 +48,12 @@ export class TaskListComponent {
 
   isOpen = signal(true);
 
-  constructor(private cdr: ChangeDetectorRef, private listService: ListService) {}
+  constructor(
+    private cdr: ChangeDetectorRef,
+    private listService: ListService,
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService
+  ) {}
 
   ngOnInit() {
     this.isOpen.set(this.taskList?.isOpen ?? true);
@@ -58,19 +70,48 @@ export class TaskListComponent {
             label: 'Excluir',
             icon: 'pi pi-trash',
             command: () => {
-              this.listService.deleteList(this.taskList!.id).subscribe({
-                next: () => {
-                  console.log('List deleted');
-                },
-                error: (error) => {
-                  console.error('Error deleting list:', error);
-                },
-              });
+              this.confirm2(event!);
             },
           },
         ],
       },
     ];
+  }
+
+  confirm2(event: Event) {
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Você tem certeza que deseja excluir este item?',
+      header: 'Confirmação de Exclusão',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass: 'p-button-danger p-button-text',
+      rejectButtonStyleClass: 'p-button-text p-button-text',
+      acceptIcon: 'none',
+      rejectIcon: 'none',
+
+      accept: () => {
+        this.messageService.add({
+          severity: 'info',
+          summary: 'Sucesso',
+          detail: 'Lista excluída',
+        });
+        this.listService.deleteList(this.taskList!.id).subscribe({
+          next: () => {
+            console.log('List deleted');
+          },
+          error: (error) => {
+            console.error('Error deleting list:', error);
+          },
+        });
+      },
+      reject: () => {
+        this.messageService.add({
+          severity: 'error',
+          summary: 'Rejected',
+          detail: 'You have rejected',
+        });
+      },
+    });
   }
 
   toggleOpen() {
